Revert "Separate Build and Serving parts of OpenAPI spec handler"
This reverts commit 0a886ffaf8b9de97ef8134a4182b719ba2c6f22f.
