+++ staging/src/k8s.io/apiserver/pkg/server/routes/openapi.go
@@ -32,16 +32,9 @@ type OpenAPI struct {
 }
 
 // Install adds the SwaggerUI webservice to the given mux.
-func (oa OpenAPI) Install(c *restful.Container, mux *mux.PathRecorderMux) *apiserveropenapi.OpenAPIService {
-	openapiSpec, err := apiserveropenapi.BuildSwaggerSpec(c.RegisteredWebServices(), oa.Config)
+func (oa OpenAPI) Install(c *restful.Container, mux *mux.PathRecorderMux) {
+	err := apiserveropenapi.RegisterOpenAPIService("/swagger.json", c.RegisteredWebServices(), oa.Config, mux)
 	if err != nil {
 		glog.Fatalf("Failed to register open api spec for root: %v", err)
-		return nil
 	}
-	service, err := apiserveropenapi.RegisterOpenAPIService(openapiSpec, "/swagger.json", mux)
-	if err != nil {
-		glog.Fatalf("Failed to register open api spec for root: %v", err)
-		return nil
-	}
-	return service
 }
